Rename vfs-severity-wizard to vfs-severity-map
