+++ packages/steiger/src/features/run-rule/prepare-vfs-for-rule-run.ts
@@ -1,6 +1,6 @@
 import { Folder } from '@steiger/types'
 
-import { getVfsWithoutOffNodes } from '../../models/vfs-severity-wizard'
+import { getVfsWithoutOffNodes } from '../../models/vfs-severity-map'
 import { GlobGroup } from '../../models/config'
 
 export function prepareVfsForRuleRun(vfs: Folder, globGroups: Array<GlobGroup>) {
